add upload image

+++ database/migrations/2019_03_06_145306_create_product_images_table.php
@@ -17,7 +17,7 @@ class CreateProductImagesTable extends Migration
         Schema::create('product_images', function (Blueprint $table) {
             $table->engine = 'InnoDB';
             $table->bigIncrements('id');
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->string('img_link');
             $table->timestamps();
         });
